10/22(일) 수정사항 반영
