+++ market/src/App.tsx
@@ -1,33 +1,5 @@
-import { useEffect, useState } from 'react';
-import axios from 'axios';
 import ProductList from './components/ProductList';
-import { product } from './types';
-
-
-function useProducts() {
-  const [data, setData] = useState<product[]>([]);
-  const [loading, setLoading] = useState(true);
-  const [error, setError] = useState<Error | null>(null);
-
-  useEffect(() => {
-    async function fetchProducts() {
-      try {
-        const { data } = await axios.get('https://fakestoreapi.com/products');
-        setData(data);
-      } catch (err) {
-        if (err instanceof Error) {
-          setError(err);
-        }
-      } finally {
-        setLoading(false);
-      }
-    }
-
-    fetchProducts();
-  }, []);
-
-  return { data, loading, error };
-}
+import useProducts from './hooks/useProducts';
 
 function App() {
   const { data, loading, error } = useProducts();
