+++ market/src/types/index.ts
@@ -1,5 +1,5 @@
-export type product = {
-  id: number;
+export type Product = {
+  id?: number;
   title: string;
   price: number;
   description: string;
